+++ Changelog.md
@@ -14,8 +14,9 @@ v2.15.0 (not published yet)
 
 - Fix misleading language name fetch from path for subtitles ([#924](/../../issues/924))
 - Add *.av1* to valid input extensions for mkvmerge ([#982](/../../issues/982))
-- Add SharpenersPack funtions
+- Add SharpenersPack funtions for dependencies
 - NVEncC: add --vpp-nnedi fields ([#962](/../../issues/962))
+- QSVEnc: add --output-depth parameter ([#984](/../../issues/984))
 - VCEEnc: Fix deprecated --vpp-resize paremeters ([#921](/../../issues/921))
 - Update tools
     - NVEnc v7.21